+++ servers/gateway/handlers/cors.go
