+++ servers/gateway/handlers/rootHandler.go
