+++ servers/gateway/scripts/gencert.sh
