+++ servers/gateway/scripts/run_local.sh
