+++ servers/gateway/tls/privkey.pem
